refactor(queue): simplify stack and extract runtime
