+++ packages/queue/src/providers/cloudflare.ts
@@ -59,10 +59,8 @@ function resolveAction(action: CloudflareQueueBatchErrorAction | void, message: 
 export function createCloudflareQueueBatchHandler<TPayload = unknown>(
   options: CloudflareQueueBatchHandlerOptions<TPayload>,
 ): (batch: CloudflareQueueMessageBatch<TPayload>) => Promise<void> {
-  const requestedConcurrency = Number(options.concurrency ?? 1)
-  const concurrency = Number.isFinite(requestedConcurrency) && requestedConcurrency > 0
-    ? Math.floor(requestedConcurrency)
-    : 1
+  const requested = Number(options.concurrency ?? 1)
+  const concurrency = Number.isFinite(requested) ? Math.max(1, Math.floor(requested)) : 1
 
   return async (batch) => {
     const messages = Array.isArray(batch?.messages) ? batch.messages : []
