+++ packages/queue/src/providers/vercel.ts
@@ -47,7 +47,7 @@ async function loadVercelQueueClient(region?: string): Promise<VercelQueueSDK> {
 }
 
 export async function createVercelQueueClient(
-  provider: VercelQueueProviderOptions & { topic?: string },
+  provider: VercelQueueProviderOptions,
 ): Promise<VercelQueueClient> {
   const topic = provider.topic
   if (!topic) {
@@ -85,7 +85,7 @@ export async function createVercelQueueClient(
 
       return {
         status: "queued",
-        messageId: result.messageId || undefined,
+        messageId: result.messageId ?? undefined,
       }
     },
     callback: client.handleCallback,
